+++ src/Products/PloneMeeting/content/category.py
@@ -141,19 +141,35 @@ class MeetingCategory(Item):
     def is_selectable(self, userId, ignore_using_groups=False):
         '''Check if category may be used :
            - enabled;
+           - used in MeetingConfig.usedMeetingAttributes or
+             MeetingConfig.usedItemAttributes;
            - current user in using_groups or current user is Manager.'''
         tool = api.portal.get_tool('portal_plonemeeting')
-        # If we have using_groups make sure userId is creator for one of it
-        selectable = self.enabled
         cfg = tool.getMeetingConfig(self)
-        # check using_groups if relevant
-        if selectable and \
-           (not ignore_using_groups and self.get_using_groups()) and \
-           not tool.isManager(realManagers=True):
-            selectable_org_uids = tool.get_selectable_orgs(cfg, user_id=userId, the_objects=False)
-            # Check intersection between self.usingGroups and org uids for which
-            # the current user is creator
-            selectable = bool(set(self.get_using_groups()).intersection(selectable_org_uids))
+        selectable = self.enabled
+        if selectable:
+            # selectable if used, may be used in an item or meeting attribute
+            cat_type = self.get_type()
+            is_used = False
+            # meeting category
+            if cat_type == 'meetingcategories':
+                is_used = 'category' in cfg.getUsedMeetingAttributes()
+            else:
+                # item category or classifier
+                if cat_type == 'categories':
+                    is_used = 'category' in cfg.getUsedItemAttributes()
+                else:
+                    is_used = 'classifier' in cfg.getUsedItemAttributes()
+            selectable = is_used
+            # If we have using_groups make sure userId is creator for one of it
+            # check using_groups if relevant
+            if selectable and \
+               (not ignore_using_groups and self.get_using_groups()) and \
+               not tool.isManager(realManagers=True):
+                selectable_org_uids = tool.get_selectable_orgs(cfg, user_id=userId, the_objects=False)
+                # Check intersection between self.usingGroups and org uids for which
+                # the current user is creator
+                selectable = bool(set(self.get_using_groups()).intersection(selectable_org_uids))
         return selectable
 
     def get_groups_in_charge(self, the_objects=False):